+++ tests/testthat/test_run_model.R
@@ -198,7 +198,8 @@ test_that(
     )
     expect_output(
       str(run_models(mod1, mod2,
-                     parameters = par1, cost = x, effect = y)),
+                     parameters = par1, cost = x, effect = y,
+                     method = "beginning")),
       '2 obs. of  5 variables:
  $ x           : num  309300 933900
  $ y           : num  283300 891300
@@ -207,8 +208,11 @@ test_that(
  $ .effect     : num  283300 891300',
       fixed = TRUE
     )
-    s_mod <- summary(run_models(mod1, mod2,
-                                parameters = par1, cost = x, effect = y))
+    s_mod <- summary(
+      run_models(
+        mod1, mod2,
+        parameters = par1, cost = x, effect = y,
+        method = "beginning"))
     expect_length(
       s_mod, 6
     )
@@ -223,7 +227,8 @@ test_that(
     )
     
     res_b <- run_models(mod1, mod2,
-                        parameters = par1, cost = x, effect = y)
+                        parameters = par1, cost = x, effect = y,
+                        method = "beginning")
     res_e <- run_models(mod1, mod2,
                         parameters = par1, cost = x, effect = y,
                         method = "end")
@@ -328,15 +333,18 @@ test_that("Discounting", {
     X2 = s4
   )
   res <- run_models(mod1, mod2, cycles = 10,
-                    parameters = par1, cost = x, effect = y)
+                    parameters = par1, cost = x, effect = y,
+                    method = "beginning")
   expect_output(
     print(res),
     "II 3292.352 4193.422 0.7851231"
   )
   res1 <- run_models(mod1, mod2, cycles = 10,
-                     parameters = par1, cost = x, effect = y)
+                     parameters = par1, cost = x, effect = y,
+                     method = "beginning")
   res2 <- run_models(mod3, mod2, cycles = 10,
-                     parameters = par1, cost = x, effect = y)
+                     parameters = par1, cost = x, effect = y,
+                     method = "beginning")
   expect_output(
     print(res1),
     "I  3144649 2942952